+++ Fallen-8/IWrite.cs
@@ -57,7 +57,7 @@ namespace NoSQL.GraphDB
         /// <param name='targetVertexId'> Target vertex identifier. </param>
         /// <param name='creationDate'> The creation date as Unix timestamp (seconds from 01/01/1971) </param>
         /// <param name='properties'> Properties. </param>
-        EdgeModel CreateEdge(Int32 sourceVertexId, UInt16 edgePropertyId, Int32 targetVertexId, UInt32 creationDate,
+        EdgeModel CreateEdge(Int64 sourceVertexId, UInt16 edgePropertyId, Int64 targetVertexId, UInt32 creationDate,
                              PropertyContainer[] properties = null);
 
         #endregion
@@ -71,7 +71,7 @@ namespace NoSQL.GraphDB
         /// <param name='graphElementId'> Graph element identifier. </param>
         /// <param name='propertyId'> Property identifier. </param>
         /// <param name='property'> The to be added property. </param>
-        Boolean TryAddProperty(Int32 graphElementId, UInt16 propertyId, Object property);
+        Boolean TryAddProperty(Int64 graphElementId, UInt16 propertyId, Object property);
 
         /// <summary>
         ///   Tries to remove a property.
@@ -79,7 +79,7 @@ namespace NoSQL.GraphDB
         /// <returns> <c>true</c> if the property was removed; otherwise, <c>false</c> . </returns>
         /// <param name='graphElementId'> Graph element identifier. </param>
         /// <param name='propertyId'> Property identifier. </param>
-        Boolean TryRemoveProperty(Int32 graphElementId, UInt16 propertyId);
+        Boolean TryRemoveProperty(Int64 graphElementId, UInt16 propertyId);
 
         #endregion
 
@@ -90,7 +90,7 @@ namespace NoSQL.GraphDB
         /// </summary>
         /// <returns> <c>true</c> if the graph element was removed; otherwise, <c>false</c> . </returns>
         /// <param name='graphElementId'> Graph element identifier. </param>
-        Boolean TryRemoveGraphElement(Int32 graphElementId);
+        Boolean TryRemoveGraphElement(Int64 graphElementId);
 
         #endregion
