+++ Test/IntegrationTests/IGraphServiceIntegrationTest.cs
@@ -1,5 +1,6 @@
 ﻿#region Usings
 
+using System;
 using System.Collections.Generic;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using NoSQL.GraphDB.Service.REST;
@@ -87,8 +88,8 @@ namespace NoSQL.GraphDB.Test
 
             IGraphService target = CreateIGraphService(); // TODO: Initialize to an appropriate value
             EdgeSpecification definition = null; // TODO: Initialize to an appropriate value
-            int expected = 0; // TODO: Initialize to an appropriate value
-            int actual;
+            Int64 expected = 0; // TODO: Initialize to an appropriate value
+            Int64 actual;
             actual = target.AddEdge(definition);
             Assert.AreEqual(expected, actual);
         }
@@ -119,8 +120,8 @@ namespace NoSQL.GraphDB.Test
 
             IGraphService target = CreateIGraphService(); // TODO: Initialize to an appropriate value
             VertexSpecification definition = null; // TODO: Initialize to an appropriate value
-            int expected = 0; // TODO: Initialize to an appropriate value
-            int actual;
+            Int64 expected = 0; // TODO: Initialize to an appropriate value
+            Int64 actual;
             actual = target.AddVertex(definition);
             Assert.AreEqual(expected, actual);
         }
@@ -265,8 +266,8 @@ namespace NoSQL.GraphDB.Test
             IGraphService target = CreateIGraphService(); // TODO: Initialize to an appropriate value
             string vertexIdentifier = string.Empty; // TODO: Initialize to an appropriate value
             string edgePropertyIdentifier = string.Empty; // TODO: Initialize to an appropriate value
-            List<int> expected = null; // TODO: Initialize to an appropriate value
-            List<int> actual;
+            List<Int64> expected = null; // TODO: Initialize to an appropriate value
+            List<Int64> actual;
             actual = target.GetIncomingEdges(vertexIdentifier, edgePropertyIdentifier);
             Assert.AreEqual(expected, actual);
         }
@@ -315,8 +316,8 @@ namespace NoSQL.GraphDB.Test
             IGraphService target = CreateIGraphService(); // TODO: Initialize to an appropriate value
             string vertexIdentifier = string.Empty; // TODO: Initialize to an appropriate value
             string edgePropertyIdentifier = string.Empty; // TODO: Initialize to an appropriate value
-            List<int> expected = null; // TODO: Initialize to an appropriate value
-            List<int> actual;
+            List<Int64> expected = null; // TODO: Initialize to an appropriate value
+            List<Int64> actual;
             actual = target.GetOutgoingEdges(vertexIdentifier, edgePropertyIdentifier);
             Assert.AreEqual(expected, actual);
         }
@@ -367,8 +368,8 @@ namespace NoSQL.GraphDB.Test
 
             IGraphService target = CreateIGraphService(); // TODO: Initialize to an appropriate value
             string edgeIdentifier = string.Empty; // TODO: Initialize to an appropriate value
-            int expected = 0; // TODO: Initialize to an appropriate value
-            int actual;
+            Int64 expected = 0; // TODO: Initialize to an appropriate value
+            Int64 actual;
             actual = target.GetSourceVertexForEdge(edgeIdentifier);
             Assert.AreEqual(expected, actual);
         }
@@ -383,8 +384,8 @@ namespace NoSQL.GraphDB.Test
 
             IGraphService target = CreateIGraphService(); // TODO: Initialize to an appropriate value
             string edgeIdentifier = string.Empty; // TODO: Initialize to an appropriate value
-            int expected = 0; // TODO: Initialize to an appropriate value
-            int actual;
+            Int64 expected = 0; // TODO: Initialize to an appropriate value
+            Int64 actual;
             actual = target.GetTargetVertexForEdge(edgeIdentifier);
             Assert.AreEqual(expected, actual);
         }
@@ -400,8 +401,8 @@ namespace NoSQL.GraphDB.Test
             IGraphService target = CreateIGraphService(); // TODO: Initialize to an appropriate value
             string propertyId = string.Empty; // TODO: Initialize to an appropriate value
             ScanSpecification definition = null; // TODO: Initialize to an appropriate value
-            IEnumerable<int> expected = null; // TODO: Initialize to an appropriate value
-            IEnumerable<int> actual;
+            IEnumerable<Int64> expected = null; // TODO: Initialize to an appropriate value
+            IEnumerable<Int64> actual;
             actual = target.GraphScan(propertyId, definition);
             Assert.AreEqual(expected, actual);
         }
@@ -416,8 +417,8 @@ namespace NoSQL.GraphDB.Test
 
             IGraphService target = CreateIGraphService(); // TODO: Initialize to an appropriate value
             IndexScanSpecification definition = null; // TODO: Initialize to an appropriate value
-            IEnumerable<int> expected = null; // TODO: Initialize to an appropriate value
-            IEnumerable<int> actual;
+            IEnumerable<Int64> expected = null; // TODO: Initialize to an appropriate value
+            IEnumerable<Int64> actual;
             actual = target.IndexScan(definition);
             Assert.AreEqual(expected, actual);
         }
@@ -432,8 +433,8 @@ namespace NoSQL.GraphDB.Test
 
             IGraphService target = CreateIGraphService(); // TODO: Initialize to an appropriate value
             RangeIndexScanSpecification definition = null; // TODO: Initialize to an appropriate value
-            IEnumerable<int> expected = null; // TODO: Initialize to an appropriate value
-            IEnumerable<int> actual;
+            IEnumerable<Int64> expected = null; // TODO: Initialize to an appropriate value
+            IEnumerable<Int64> actual;
             actual = target.RangeIndexScan(definition);
             Assert.AreEqual(expected, actual);
         }
@@ -480,8 +481,8 @@ namespace NoSQL.GraphDB.Test
 
             IGraphService target = CreateIGraphService(); // TODO: Initialize to an appropriate value
             SearchDistanceSpecification definition = null; // TODO: Initialize to an appropriate value
-            IEnumerable<int> expected = null; // TODO: Initialize to an appropriate value
-            IEnumerable<int> actual;
+            IEnumerable<Int64> expected = null; // TODO: Initialize to an appropriate value
+            IEnumerable<Int64> actual;
             actual = target.SpatialIndexScanSearchDistance(definition);
             Assert.AreEqual(expected, actual);
         }
